kafka-cv/requirements.txt
Flask-SocketIO==5.0.1
requests==2.25.0
kafka-python==2.0.2
msgpack-python==0.5.6
numpy==1.18.5

tensorflow==2.3.0
opencv-python-headless==4.4.0.46
lxml==4.6.2
tqdm==4.55.0
absl-py==0.11.0
easydict==1.9
matplotlib
pillow